Edited

+++ src/controllers/admin.js
@@ -32,7 +32,7 @@ module.exports.PostAddProduct = async (req, res, next) => {
 
 module.exports.GetProducts = async (req, res, next) => {
   try {
-    const products = await Product.find();
+    const products = await Product.find({ owner_id: req.user._id });
     res.status(200);
     res.render("pages/admin/products", {
       pageTitle: "Admin Products",
@@ -51,10 +51,11 @@ module.exports.GetProductToEdit = async (req, res, next) => {
   const { productId } = req.params;
   try {
     const product = await Product.findById(productId);
-    if (!product) {
-      res.redirect("/");
+    if (!product || product.owner_id != req.user._id) {
+      res.redirect("/admin/products");
+      return;
     }
-    const { edit } = req.query;
+    // const { edit } = req.query;
     res.render("pages/admin/edit-product", {
       pageTitle: `${product ? "Edit Product" : "error"}`,
       product: product,
@@ -73,6 +74,10 @@ module.exports.PostUpdateProduct = async (req, res, next) => {
   const { title, image_url, price, description } = req.body;
   try {
     let product = await Product.findById(productId);
+    if (product.owner_id != req.user._id) {
+      res.redirect("/admin/products");
+      return;
+    }
     product.title = title;
     product.image_url = image_url;
     product.price = price;
@@ -89,6 +94,11 @@ module.exports.PostUpdateProduct = async (req, res, next) => {
 module.exports.PostDeleteProduct = async (req, res, next) => {
   const { productId } = req.params;
   try {
+    const product = await Product.findById(productId);
+    if (product.owner_id != req.user._id) {
+      res.redirect("/admin/products");
+      return;
+    }
     await Product.findByIdAndDelete(productId);
     res.redirect("/admin/products");
   } catch (error) {
